Fix ARABIC LETTER HIGH HAMZA (U+0674)
It is a spacing not a mark, its size should be that of ARABIC LETTER
HAMZA (U+0621) but raised above baseline. I made it a component of the
regular hamza, remove the erroneous mkmk anchors and raised it above
baseline.

I also fixed the deprecated characters that use it like ARABIC LETTER
HIGH HAMZAH WAW (U+0676), but only where it makes sense (i.e. isolated
and initial forms, since medial or final forms of these characters are
not attested).