Naskh: Sort glyphs in a more structured way
Keep base (dotless) glyphs at top, move placeholder glyphs (that we
decompose) to a made up category, and sort ligatures in the same order
as their components.

This mainly keep important glyphs at the top, so I don’t have to wade
through glyphs that I rarely touch, and to make it easy to locate a
specific glyph by eye.