Naskh: Remove a misguided substitution
High hamza is non-joining letter and not a mark, so an initial lam would
never come before it (it would be isolated lam), so this substitution is
not used. This is a left over when high hamza was confused for being a
variant from hamza above and was GDEF classified as a mark, and seems to
have been meant to support the Quranic hamza floating between lam and
alef, which we already support the proper way to encode it: lam +
tatweel + hamza above + alef.