Naskh: sort out stop sign positioning
These are the top most level of Arabic marks, so they don’t need mkmk
anchors on top of them.

Furthermore, it is preferable that they have consistent vertical
position, and they usually apply to the end of a word, not the specific
letter they are above, so they don’t typically need mark anchors either,
and instead should be placed high enough to be above any glyph. However,
some orthographies put them on space or end of ayah mark, where they
should not be too high, so we now use a dedicated anchor that are used
only with those.

seenabove-ar is a special case, as though it us usually used as stop
sign, it is also used as mark above medial sad in some Qur’anic words,
so we also used a dictated anchor for that.