Naskh: Remove effectively unused hamza mark ligatures
The mark ligature lookup is comes before the decomposition lookup, so
these ligatures are not used for decomposed letters with hmaza, and will
be only used when hamza is used as a combining mark which rather rare,
so they are effectively unused. They are also superfluous since mark to
mark positioning does the job.