Naskh: Add more Sindhi local forms
Add an alternate form of meem-ar and meem-ar.fina that is preferred in
Sindhi, and an alternate form of sindhipostpositionmen-ar that uses the
alternate meem (we could have changed the default glyph since it is a
Sindhi-specific character, but for sake of consistency we keep the
default form matching the default form of meem).

Sindhi also prefers reversed comma and semicolon, so we substitute
comma-ar and semicolon-ar as well.