Add anchor to space and thinspace glyphs
They are often used as bases for standalone marks, and it seems without
anchors Chrome will badly position the marks on them (I think it might
be splitting the text at the space glyph and partially shaping it).

For example, the word فَٱدَّ ٰرَءٖۡتُمۡ has the small alef above mispositioned in
Chrome with Noto Sans Arabic.

FontBakery also complains that we are not attaching marks to the space,
and this change should fix it as well.